Move age-based cleanup from acquire() to release()
Cleanup expired stacks in release() instead of acquire() to avoid
the situation where we remove the last stack and then immediately
need to allocate a new one. This is more efficient - we clean up
expired stacks when adding a new stack to the pool, ensuring we
never deplete the pool unnecessarily.